+++ packages/components/src/index.ts
@@ -13,3 +13,4 @@ export { MinisNavigationItem } from './components/navigation/navigation-item.js'
 export { MinisCardGrid } from './components/card-grid/card-grid.js';
 export { MinisContainer } from './components/container/container.js';
 export { MinisTopbar } from './components/topbar/topbar.js';
+export { MinisCheckbox } from './components/checkbox/checkbox.js';